+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- `doCleanup` now takes a variadic amount of tasks
 - Reorganized the Fusion namespace
   - `Contextual` has been moved to State
   - `Safe` has been moved to Components